+++ db/events.php
@@ -28,11 +28,11 @@ $observers = array(
     array(
         'eventname'   => '\core\event\course_created',
         'includefile' => '/local/resort_courses/locallib.php',
-        'callback'    => 'resort_courses',
+        'callback'    => 'resort_course_eventhandler',
     ),
     array(
         'eventname'   => '\core\event\course_updated',
         'includefile' => '/local/resort_courses/locallib.php',
-        'callback'    => 'resort_courses',
+        'callback'    => 'resort_course_eventhandler',
     ),
 );